requirements.txt
Flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
pdfplumber==0.10.3
Werkzeug==3.0.1
gunicorn==21.2.0
